+++ app/views/admin/lead_providers/lead_provider_users/edit.html.erb
@@ -1,7 +1,7 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
     <h1 class="govuk-heading-l">Edit user details</h1>
-    <%= form_for(@lead_provider_user, url: admin_lead_provider_user_path(@lead_provider, @lead_provider_user)) do |f| %>
+    <%= form_for(@lead_provider_user, url: admin_lead_provider_user_path(@lead_provider_user)) do |f| %>
       <%= f.govuk_error_summary %>
       <%= f.govuk_text_field(
             :full_name,
