--- app/views/admin/suppliers/lead_provider_users/_form_fields.html.erb
@@ -1,9 +1,0 @@
-<%= f.govuk_text_field(
-      :full_name,
-      label: { text: "Full name" })
-%>
-
-<%= f.govuk_email_field(
-      :email,
-      label: { text: "Email" })
-%>
