+++ spec/requests/admin/lead_providers/lead_provider_users_spec.rb
@@ -2,9 +2,8 @@
 
 require "rails_helper"
 
-RSpec.describe "Admin::LeadProviderUsers", type: :request do
+RSpec.describe "Admin::LeadProviders::LeadProviderUsers", type: :request do
   let(:lead_provider_profile) { create(:lead_provider_profile) }
-  let(:lead_provider) { lead_provider_profile.lead_provider }
   let(:lead_provider_user) { lead_provider_profile.user }
 
   before do
@@ -12,12 +11,12 @@ RSpec.describe "Admin::LeadProviderUsers", type: :request do
     sign_in admin_user
   end
 
-  describe "GET /admin/lead_providers/:id/users/:id/edit" do
+  describe "GET /admin/lead_providers/users/:id/edit" do
     it "renders the index template" do
-      get "/admin/lead-providers/#{lead_provider.id}/users/#{lead_provider_user.id}/edit"
+      get "/admin/lead-providers/users/#{lead_provider_user.id}/edit"
 
       expect(response.body).to include("Edit user details")
-      expect(response).to render_template("admin/lead_provider_users/edit")
+      expect(response).to render_template("admin/lead_providers/lead_provider_users/edit")
     end
   end
 
@@ -25,7 +24,7 @@ RSpec.describe "Admin::LeadProviderUsers", type: :request do
     let(:email) { "user@example.com" }
 
     it "updates the user and redirects to users page" do
-      patch "/admin/lead-providers/#{lead_provider.id}/users/#{lead_provider_user.id}", params: {
+      patch "/admin/lead-providers/users/#{lead_provider_user.id}", params: {
         user: { email: email },
       }
 
@@ -36,12 +35,12 @@ RSpec.describe "Admin::LeadProviderUsers", type: :request do
 
     context "when the user params are invalid" do
       it "renders error messages" do
-        patch "/admin/lead-providers/#{lead_provider.id}/users/#{lead_provider_user.id}", params: {
+        patch "/admin/lead-providers/users/#{lead_provider_user.id}", params: {
           user: { email: nil },
         }
 
         expect(response.body).to include("Enter an email")
-        expect(response).to render_template("admin/lead_provider_users/edit")
+        expect(response).to render_template("admin/lead_providers/lead_provider_users/edit")
       end
     end
   end
